pyproject.toml
[project]
name = "TaxChat"
version = "0.1.0"
description = "TaxChat is a chatbot that can answer questions about tax law."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[openai]>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "tiktoken>=0.7.0",
    "langchain-text-splitters>=1.0.0",
    "bs4>=0.0.2",
    "langchain-chroma>=1.0.0",
    "streamlit>=1.28.0",
    "rich>=13.0.0",
]
